pyproject.toml
[project]
name = "st-hello-world"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.10.4",
    "polars>=1.26.0",
    "pip>=25.0.1",
    "databricks-sql-connector>=4.0.3",
    "streamlit>=1.45.0",
]

[dependency-groups]
dev = [
    "databricks-connect==16.2.0",
    "databricks-dlt>=0.3.0",
    "databricks-sdk[notebook]>=0.49.0",
    "ipykernel>=6.29.5",
    "pip>=25.0.1",
    "plotly>=6.0.1",
    "pytest>=8.3.4",
    "python-dotenv>=1.0.1",
    "ruff>=0.8.4",
    "polars>=1.26.0",
    "pip>=25.0.1",
    "databricks-sql-connector>=4.0.3",
    "streamlit>=1.45.0",
]
